# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Authentication and Authorization API",
        "description": "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).For this sample, you can use the api key `special-key` to test the authorization     filters.",
        "termsOfService": "http://Lanthanion.com",
        "license": {
            "name": "Apache 2.0",
            "url": "http://springdoc.org"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/users/getparticularuser": {
            "post": {
                "tags": [
                    "users-controller"
                ],
                "operationId": "getParticularUser",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/GetParticularUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralUsersResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/getAllUsers": {
            "post": {
                "tags": [
                    "users-controller"
                ],
                "operationId": "getAllUsers",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/GeneralPagedRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralPagedUsersResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/deleteuser": {
            "delete": {
                "tags": [
                    "users-controller"
                ],
                "operationId": "deleteUser",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/DeleteUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/getuserbyusertype": {
            "post": {
                "tags": [
                    "users-controller"
                ],
                "operationId": "getUsersByUserType",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/GetUsersByUserTypeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralPagedUsersResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/updateuserdetails": {
            "put": {
                "tags": [
                    "users-controller"
                ],
                "operationId": "updateUserDetails",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserDetailsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralUsersResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/resetuserpassword": {
            "post": {
                "tags": [
                    "users-controller"
                ],
                "operationId": "resetUserPassword",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetUserPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "GeneralPagedRequest": {
                "required": [
                    "isDeleted",
                    "items",
                    "page"
                ],
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "items": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "isDeleted": {
                        "type": "boolean"
                    }
                }
            },
...
            "Users": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "isLoggedIn": {
                        "type": "boolean"
                    },
                    "userType": {
                        "type": "string",
                        "enum": [
                            "ADVERTISER",
                            "AD_OWNER",
                            "AFFILIATE",
                            "ADMIN"
                        ]
                    },
                    "isDeleted": {
                        "type": "boolean"
                    },
                    "deletedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdBy": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "updatedBy": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "GeneralResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "GeneralPagedUsersResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Users"
                        }
                    }
                }
            },
...
            "GetParticularUserRequest": {
                "required": [
                    "userId"
                ],
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "GeneralUsersResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Users"
                    }
                }
            },
            "DeleteUserRequest": {
                "required": [
                    "userId"
                ],
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "GetUsersByUserTypeRequest": {
                "required": [
                    "isDeleted",
                    "items",
                    "page",
                    "userType"
                ],
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "items": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "isDeleted": {
                        "type": "boolean"
                    },
                    "userType": {
                        "type": "string",
                        "enum": [
                            "ADVERTISER",
                            "AD_OWNER",
                            "AFFILIATE",
                            "ADMIN"
                        ]
                    }
                }
            },
            "UpdateUserAddress": {
                "type": "object",
                "properties": {
                    "country": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "street": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "postalCode": {
                        "type": "string"
                    }
                }
            },
            "UpdateUserDetailsRequest": {
                "required": [
                    "userId"
                ],
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "email": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "roleId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "userType": {
                        "type": "string",
                        "enum": [
                            "ADVERTISER",
                            "AD_OWNER",
                            "AFFILIATE",
                            "ADMIN"
                        ]
                    },
                    "password": {
                        "type": "string"
                    },
                    "updateUserAddress": {
                        "$ref": "#/components/schemas/UpdateUserAddress"
                    }
                }
            },
            "ResetUserPasswordRequest": {
                "required": [
                    "password",
                    "userId"
                ],
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "API KEY / API SECRET": {
                "type": "apiKey",
                "scheme": "apiKey"
            }
        }
    }
}